Only show progress bar if verbosity level is DEBUG
This change will skip the rendering of the connectivity computation
progress bar if the logging level is not DEBUG. This is in line with
MNE-Python, where progress bars are not shown at INFO or higher logging
levels. Rendering the progress bar regardless of logging levels has the
potential to cause unnecessary clutter in users' log files.